Enable opening topologies without "gui" field
IMUNES can open the topology configure without any GUI fields - if
opening in 'interactive/GUI' mode, it will create random coordinates
for each node and rearrange them for a second after opening it. All
nodes will be created on the same (the only) canvas and all the other
GUI stuff will be set to default values.