Working PoC for remote execution
Connecting to a remote Linux/FreeBSD machine and running commands works.
It's a bit slow and inefficient, but it works.

There are two ways for connecting: ssh and socat. Both need passwordless
SSH access and sudo installed on REMOTE machine. It doesn't have to be
sudo, but it's currently hardcoded as sudo in imunes.tcl
(rescalation_comm variable). To run wireshark/tcpdump on local machine,
it may be required to also set the escalation_comm variable to whatever
you use LOCALLY.

Requirements:
 - local requirements:
  - IMUNES installed
  - ssh access to remote using pubkey auth
  - scp installed (hardcoded for now)
  - $rescalation_comm command installed (default is sudo)
 - remote requirements:
  - IMUNES installed - not really!

1. SSH
 - remote requirements:
  - IMUNES installed - not really!
  - user with ssh access using pubkey auth
 - Usage example
    $ ./imunes -r 10.0.0.20

2. socat
 - local requirements:
  - socat installed
 - remote requirements:
  - IMUNES installed - not really!
  - socat installed and listening:
   $ sudo -i socat TCP-LISTEN:1234,reuseaddr,fork EXEC:/bin/sh

Usage example
$ ./imunes -r user@10.0.0.20		# ssh
$ ./imunes -r socat:10.0.0.20:1234	# socat