Enable experimental remote execution
This makes it possible to connect to a remote Linux/FreeBSD machine and
run IMUNES commands on that machine.
It's a bit slow and inefficient, but it works.

To connect, you need passwordless SSH access and some escalation command
installed on the REMOTE machine. The default is 'sudo -i' but it's
possible to change it by providing -r argument as
'remote_host%rescalation_comm'.

Additionally, if your local wireshark needs root access, provide -r
argument as 'escalation_comm/remote_host'.

NOTE: use + instead of spaces for arguments with multiple words, check
usage examples at the end.

The number of parallel commands sent to the remote server is calculated
by min(remote number of CPUs, remote_max_sessions), where
remote_max_sessions can be changed by providing the -rm argument.

NOTE: the default value of MaxSessions on SSH servers is 10, the default
in IMUNES is set to 8.

Requirements:
 - local requirements:
  - IMUNES installed (no docker/vroot required)
  - ssh access to remote using pubkey auth
 - remote requirements:
  - Linux: installed iproute2, docker, imunes/template docker image and
    IMUNES directories (/var/run/imunes /var/imunes/ /var/log/imunes)
  - FreeBSD: installed vroot and IMUNES directories
  - $rescalation_comm command installed (default is sudo -i)
  - user with ssh access using pubkey auth

Usage example
$ ./imunes -r user@10.0.0.20	# regular SSH
$ ./imunes -r test		# .ssh/config host entry 'test'
$ ./imunes -r sudo/10.0.0.20	# use 'sudo' for local wireshark
$ ./imunes -r 10.0.0.20%sudo+-i	# use 'sudo -i' for remote cmds