Extend IMUNES rc config files
Read existing config files, in this order:
 - ./.imunesrc
 - /etc/imunes/config
 - $HOME/.imunes.rc if it exists, otherwise $XDG_CONFIG_HOME/imunes/config
 - ./.imunes.rc

After that, read /etc/imunes/override which overrides any previously set
options.

All the configuration files are JSON files, except for .imunesrc, which
is a TCL file which will be sourced. It is kept for compatibility
reasons.

NOTE: If $XDG_CONFIG_HOME is either not set or empty, a default of
$HOME/.config is used

Check config.example for the list of options.