Mark entries as set/overridden by custom config
If the menu entry is green, this means that some rc config file has this
option set.

If the menu entry is gray, this means that some rc config file has this
option in its 'custom_override' variable. If the option is not set in
any of the files, it will always take the default value.